// src/app/core/models/alumno.model.ts
export interface Alumno {
  id_alumno?: number;
  codigo: string;
  dni: string;
  nombres: string;
  apellidos: string;
  nivel: number;
  nivel_nombre?: string;
  sexo: string;
  fecha_nacimiento: string;
  grado: number;
  grado_nombre?: string;
  seccion: string;
  domicilio_actual: string;
  referencia: string;
  tipo_seguro: string;
  discapacidad: boolean
}